refactor(stats): make score a resource
