+++ src/stats/resources.rs
@@ -2,6 +2,10 @@ use std::time::Duration;
 
 use bevy::prelude::*;
 
+#[derive(Debug, Default, Reflect, Resource, Deref, DerefMut)]
+#[reflect(Resource)]
+pub struct Score(pub usize);
+
 #[derive(Debug, Reflect, Resource, Deref, DerefMut)]
 #[reflect(Resource)]
 pub struct TotalTimer(pub Timer);
